1.bug修改 <br> 2.功能新增與調整 <br>

+++ src/main/java/com/blog/service/impl/SubscriptionServiceImpl.java
@@ -18,19 +18,23 @@ import com.blog.service.UserService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
+import org.springframework.data.redis.core.StringRedisTemplate;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.concurrent.TimeUnit;
 
 @Slf4j
 @Service
 @RequiredArgsConstructor
-@Transactional
+@Transactional(rollbackFor = Exception.class)
 public class SubscriptionServiceImpl implements SubscriptionService {
 
     private final SubscriptionPoRepository subscriptionPoRepository;
     private final UserPoRepository userPoRepository;
+
+    private final StringRedisTemplate stringRedisTemplate;
 
     /**
      * 訂閱文章
@@ -46,6 +50,11 @@ public class SubscriptionServiceImpl implements SubscriptionService {
         SubscriptionPo subscriptionPo = SubscriptionPoMapper.INSTANCE.toPo(subscriptionDto);
         UserPo userPo = userPoRepository.findByUserName(username).orElseThrow(() -> new ResourceNotFoundException("找不到使用者"));
         subscriptionPo.setUser(userPo);
+        // 防止二次訂閱
+        if(Boolean.FALSE.equals(stringRedisTemplate.hasKey("bookmark" + username + "_" + postId))){
+            subscriptionPoRepository.save(subscriptionPo);
+            stringRedisTemplate.opsForValue().set("bookmark" + username + "_" + postId, "true");
+        }
         subscriptionPoRepository.saveAndFlush(subscriptionPo);
         return "訂閱成功";
     }
@@ -57,8 +66,12 @@ public class SubscriptionServiceImpl implements SubscriptionService {
      * @param postId 被取消訂閱的文章id
      */
     @Override
-    @Transactional(rollbackFor = Exception.class)
     public String unSubscribe(String username, Long postId) {
+        // 防止二次取消
+        if(Boolean.FALSE.equals(stringRedisTemplate.hasKey("bookmark" + username + "_" + postId))){
+            return "已取消訂閱、請勿重複操作";
+        }
+        stringRedisTemplate.delete("bookmark" + username + "_" + postId);
         subscriptionPoRepository.deleteByUsername(username);
         return "取消訂閱成功";
     }
@@ -71,8 +84,8 @@ public class SubscriptionServiceImpl implements SubscriptionService {
 
     @Override
     public String checkSubscription(String username, Long postId) {
-        if(subscriptionPoRepository.existsByUsername(username) > 0){
-           return "已訂閱";
+        if(Boolean.TRUE.equals(stringRedisTemplate.hasKey("bookmark" + username + "_" + postId))){
+            return "已訂閱";
         }
         return "未訂閱";
     }
